Merge current development build to reflect newest release version

Former-commit-id: ce53205b5611026df99ed1cea099db0f2862d3e0 [formerly a8e8511c55ed05dbcbc1b39228507d38d9256541] [formerly 093ed811f41cc1b7690f2e83d38769399f5beaa2 [formerly 06da42e9b6d6e0e6dfce3a219aecff4231cbf7e9]]
Former-commit-id: 066d8db541c3b5a696a5e080131ffc90c8554e87 [formerly 779cd576b7d59043af2faee6a79a49fa1ac28cf8]
Former-commit-id: b6064fa2e1a25532b67d80731791066a62783e6e

+++ src/main/java/asl/sensor/utils/ReportingUtils.java
@@ -7,7 +7,6 @@ import java.awt.image.BufferedImage;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
-
 import org.apache.pdfbox.pdmodel.PDDocument;
 import org.apache.pdfbox.pdmodel.PDPage;
 import org.apache.pdfbox.pdmodel.PDPageContentStream;
@@ -35,68 +34,68 @@ public class ReportingUtils {
    * @param pdf PDF document to append the page onto
    * @param bis Series of buffered images to merge onto a single page
    */
-  public static void 
+  public static void
   bufferedImagesToPDFPage(PDDocument pdf, BufferedImage... bis) {
     BufferedImage toPDF = mergeBufferedImages(bis);
     bufferedImageToPDFPage(toPDF, pdf);
   }
-  
+
   /**
    * Add a buffered image to a PDDocument page
    * @param bi BufferedImage to be added to PDF
    * @param pdf PDF to have BufferedImage appended to
    */
-  public static void 
+  public static void
   bufferedImageToPDFPage(BufferedImage bi, PDDocument pdf) {
-    
-    PDRectangle rec = 
-        new PDRectangle( (float) bi.getWidth(), 
-                         (float) bi.getHeight() );
+
+    PDRectangle rec =
+        new PDRectangle( bi.getWidth(),
+                         bi.getHeight() );
     PDPage page = new PDPage(rec);
-    
+
     try {
-      PDImageXObject  pdImageXObject = 
+      PDImageXObject  pdImageXObject =
           LosslessFactory.createFromImage(pdf, bi);
       pdf.addPage(page);
-      PDPageContentStream contentStream = 
-          new PDPageContentStream(pdf, page, 
-                                  PDPageContentStream.AppendMode.OVERWRITE, 
+      PDPageContentStream contentStream =
+          new PDPageContentStream(pdf, page,
+                                  PDPageContentStream.AppendMode.OVERWRITE,
                                   true, false);
 
-      contentStream.drawImage( pdImageXObject, 0, 0, 
+      contentStream.drawImage( pdImageXObject, 0, 0,
           bi.getWidth(), bi.getHeight() );
       contentStream.close();
-      
+
     } catch (IOException e) {
       e.printStackTrace();
-    } 
+    }
 
     return;
-    
+
   }
-  
+
   /**
    * Converts a series of charts into a buffered image. Each chart has the
    * dimensions given as the width and height parameters, and so the resulting
    * image has width given by that parameter and height equal to height
-   * multiplied by the number of charts passed in 
+   * multiplied by the number of charts passed in
    * (that is, the charts are concatenated vertically)
    * @param width width of each chart plot
    * @param height height of each chart plot
    * @param jfcs series of charts to be plotted in
    * @return buffered image consisting of the concatenation of the given charts
    */
-  public static BufferedImage 
+  public static BufferedImage
   chartsToImage(int width, int height, JFreeChart... jfcs) {
-    
+
     BufferedImage[] bis = new BufferedImage[jfcs.length];
-    
+
     for (int i = 0; i < jfcs.length; ++i) {
       ChartPanel cp = new ChartPanel(jfcs[i]);
       cp.setSize( new Dimension(width, height) );
       BufferedImage temp = new BufferedImage(
-          (int) cp.getWidth(),
-          (int) cp.getHeight(),
+          cp.getWidth(),
+          cp.getHeight(),
           BufferedImage.TYPE_INT_ARGB);
       Graphics2D g = temp.createGraphics();
       g = temp.createGraphics();
@@ -104,10 +103,10 @@ public class ReportingUtils {
       g.dispose();
       bis[i] = temp;
     }
-    
+
     return mergeBufferedImages(bis);
   }
-  
+
   /**
    * Create a list of buffered images from a series of charts, with a specified
    * number of charts included on each image. This is used to write the charts
@@ -116,22 +115,22 @@ public class ReportingUtils {
    * @param width Width to set each chart's output image
    * @param height Height to set each chart's output image
    * @param charts List of charts to be compiled into images
-   * @return A list of buffered images with no more than perImg plots in 
+   * @return A list of buffered images with no more than perImg plots in
    * each image
    */
   public static BufferedImage[]
   chartsToImageList(int perImg, int width, int height, JFreeChart... charts) {
-    
+
     List<BufferedImage> imageList = new ArrayList<BufferedImage>();
     int totalNumber = charts.length;
-    
+
     if (totalNumber < perImg) {
       // if we can fit them all on a single page, then we'll do so
       imageList.add( chartsToImage(width, height, charts) );
       return imageList.toArray( new BufferedImage[]{} );
     }
-    
-    // want to keep all charts the same size; 
+
+    // want to keep all charts the same size;
     // if we can't fit them all on a single page, how many pages will have
     // complete charts?
     int numFilledPages = totalNumber / perImg;
@@ -140,7 +139,7 @@ public class ReportingUtils {
     // how many chart-size blank spaces to keep plots on last page same size
     int spacerCount = perImg - lastPageChartCount;
     // Note that the above variable is not used if lastPageChartCount is zero
-    
+
     // handle all the pages with complete data here
     for (int i = 0; i < numFilledPages; ++i) {
       JFreeChart[] onOnePage = new JFreeChart[perImg];
@@ -149,7 +148,7 @@ public class ReportingUtils {
       }
       imageList.add( chartsToImage(width, height, onOnePage) );
     }
-    
+
     // special case for a non-evenly dividing plot series
     if (lastPageChartCount != 0) {
       int lastIndex = numFilledPages * perImg;
@@ -161,10 +160,10 @@ public class ReportingUtils {
       BufferedImage space = createWhitespace(width, height * spacerCount);
       imageList.add( mergeBufferedImages(lastPageImage, space) );
     }
-    
+
     return imageList.toArray( new BufferedImage[]{} );
   }
-  
+
   /**
    * Takes in a series of charts and produces a PDF page of those charts.
    * For more details on this method, see the chartsToImage function, which
@@ -174,17 +173,17 @@ public class ReportingUtils {
    * @param pdf PDF document to have the data appended to
    * @param jfcs series of charts to place in the PDF
    */
-  public static void 
+  public static void
   chartsToPDFPage(int width, int height, PDDocument pdf, JFreeChart... jfcs) {
-    
+
     BufferedImage bi = chartsToImage(width, height, jfcs);
     bufferedImageToPDFPage(bi, pdf);
     return;
-    
+
   }
-  
+
   public static BufferedImage createWhitespace(int width, int height) {
-    BufferedImage out = new BufferedImage(width, height, 
+    BufferedImage out = new BufferedImage(width, height,
         BufferedImage.TYPE_INT_RGB);
     Graphics2D g = out.createGraphics();
 
@@ -193,10 +192,10 @@ public class ReportingUtils {
     g.dispose();
     return out;
   }
-  
+
   /**
    * Writes multiple pages of charts to a PDF file, with the number of charts
-   * to display per page set according to a parameter 
+   * to display per page set according to a parameter
    * @param perPage Number of charts to put in a page at a time
    * @param width Width of each chart to write to file
    * @param height Height of each chart to write to file
@@ -206,12 +205,12 @@ public class ReportingUtils {
   public static void
   groupChartsToPDFPages(int perPage, int width, int height,
       PDDocument pdf, JFreeChart... charts) {
-    
-    imageListToPDFPages( pdf, 
+
+    imageListToPDFPages( pdf,
         chartsToImageList(perPage, width, height, charts) );
-    
+
   }
-  
+
   /**
    * Write a list of images to a pdf document, each image its own page
    * @param pdf PDF document to write to
@@ -224,16 +223,16 @@ public class ReportingUtils {
       bufferedImageToPDFPage(bi, pdf);
     }
   }
-  
+
   /**
    * Utility function to combine a series of buffered images into a single
-   * buffered image. Images are concatenated vertically and centered 
+   * buffered image. Images are concatenated vertically and centered
    * horizontally into an image as wide as the widest passed-in image
    * @param bis Buffered images to send in
    * @return Single concatenated buffered image
    */
   public static BufferedImage mergeBufferedImages(BufferedImage... bis) {
-   
+
     int maxWidth = 0;
     int totalHeight = 0;
     for (BufferedImage bi : bis) {
@@ -242,11 +241,11 @@ public class ReportingUtils {
       }
       totalHeight += bi.getHeight();
     }
-    
-    BufferedImage out = 
+
+    BufferedImage out =
         new BufferedImage(maxWidth, totalHeight, BufferedImage.TYPE_INT_RGB);
     Graphics2D g = out.createGraphics();
-    
+
     int heightIndex = 0;
     for (BufferedImage bi : bis) {
       int centeringOffset = 0; // need to center the component?
@@ -256,11 +255,11 @@ public class ReportingUtils {
       g.drawImage(bi, null, centeringOffset, heightIndex);
       heightIndex += bi.getHeight();
     }
-    
+
     return out;
-    
+
   }
-  
+
   /**
    * Add pages to a PDF document consisting of textual data with a series of
    * strings, where each string is written to a separate page
@@ -269,25 +268,25 @@ public class ReportingUtils {
    */
   public static void
   textListToPDFPages(PDDocument pdf, String... toWrite) {
-    
+
     for (String onePage : toWrite) {
       textToPDFPage(onePage, pdf);
     }
-    
+
   }
-  
-  
+
+
   /**
    * Add a page to a PDF document consisting of textual data
    * @param toWrite String to add to a new PDF page
    * @param pdf Document to append the page to
    */
   public static void textToPDFPage(String toWrite, PDDocument pdf) {
-    
+
     if ( toWrite.length() == 0 ) {
       return;
     }
-    
+
       PDPage page = new PDPage();
       pdf.addPage(page);
 
@@ -348,7 +347,7 @@ public class ReportingUtils {
           contentStream.showText(line);
           contentStream.newLineAtOffset(0, -leading);
         }
-        contentStream.endText(); 
+        contentStream.endText();
         contentStream.close();
       } catch (IOException e) {
         e.printStackTrace();
